feat: move SearchBar to separate file in components, update useSearchInput
